Use 'tools.go' for necessary external tools
See https://golang.org/issue/48429 for upstream discussion on how to
track tool dependencies.

For a library, there are some tradeoffs in using the tools.go approach,
but since this is a standalone project, it seems like a reasoanble
approach to take.

This updates our Makefile and tools.go to track each binary dependency
using `go.mod`.

Notably, this also adds compatibility with go1.18 (coming next)